requirements.txt
lxml~=4.6.3
beautifulsoup4~=4.9.3
requests~=2.25.1
pandas~=1.2.4
selenium~=3.141.0
insta-scrape~=2.1.2
Pillow~=8.2.0